feat: enhance trump declaration system with AI integration and UI improvements
## Summary
- Re-enable AI trump declaration system with race condition protection
- Refine expandable trump declaration UI with context-aware text
- Add current player indicator during progressive dealing
- Consolidate hook architecture for better maintainability

## Key Features
- **AI Trump Declarations**: Sophisticated AI strategy with hand quality analysis, timing optimization, and override logic
- **Enhanced UI Text**: Dynamic titles ("Final Declaration Opportunity" vs "Final Declaration Info"), context-aware no-options text, and universal trump status format
- **Current Player Indicator**: Real-time "→ Player" display showing whose turn during dealing
- **Race Condition Protection**: AI-first priority order with state validation to prevent simultaneous declaration conflicts
- **Consolidated Architecture**: Single `useProgressiveDealing` hook replaces multiple interdependent hooks

## Technical Improvements
- Immediate resume approach for AI declarations to prevent dealing pause issues
- Context-aware text that works for both final and mid-dealing scenarios
- Better styling with proper spacing and visual hierarchy for player indicators
- Comprehensive error handling and state validation
