+++ src/hooks/useAITurns.ts
@@ -1,6 +1,6 @@
-import { useState, useEffect, useRef, useCallback } from "react";
-import { GameState, Card, GamePhase } from "../types";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { getAIMoveWithErrorHandling } from "../game/gamePlayManager";
+import { Card, GamePhase, GameState } from "../types";
 import { AI_MOVE_DELAY } from "../utils/gameTimings";
 
 type ProcessPlayFn = (cards: Card[]) => void;
